Introduction of credit invoices! (IP-111)

+++ application/modules/invoices/views/partial_item_table.php
@@ -17,8 +17,7 @@
 
         <tr id="new_item" style="display: none;">
             <td class="td-text">
-                <input type="hidden" name="invoice_id"
-                       value="<?php echo $invoice_id; ?>">
+                <input type="hidden" name="invoice_id" value="<?php echo $invoice_id; ?>">
                 <input type="hidden" name="item_id" value="" class="form-control">
                 <input type="text" name="item_name" class="lookup-item-name form-control"
                        data-typeahead=""><br>
@@ -58,29 +57,36 @@
         <?php foreach ($items as $item) { ?>
             <tr class="item">
                 <td class="td-text">
-                    <input type="hidden" name="invoice_id" value="<?php echo $invoice_id; ?>">
-                    <input type="hidden" name="item_id" value="<?php echo $item->item_id; ?>">
+                    <input type="hidden" name="invoice_id" value="<?php echo $invoice_id; ?>"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
+                    <input type="hidden" name="item_id" value="<?php echo $item->item_id; ?>"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
                     <input type="text" name="item_name" class="form-control"
-                           value="<?php echo $item->item_name; ?>">
+                           value="<?php echo $item->item_name; ?>"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
                 </td>
 
                 <td class="td-textarea">
-                    <textarea name="item_description"
-                              class="form-control"><?php echo $item->item_description; ?></textarea>
+                    <textarea name="item_description" class="form-control"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>
+                        ><?php echo $item->item_description; ?></textarea>
                 </td>
 
                 <td class="td-amount">
                     <input type="text" name="item_quantity" class="input-sm form-control"
-                           value="<?php echo format_amount($item->item_quantity * $invoice->invoice_sign); ?>">
+                           value="<?php echo format_amount($item->item_quantity * $invoice->invoice_sign); ?>"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
                 </td>
 
                 <td class="td-amount">
                     <input type="text" name="item_price" class="input-sm form-control"
-                           value="<?php echo format_amount($item->item_price * $invoice->invoice_sign); ?>">
+                           value="<?php echo format_amount($item->item_price * $invoice->invoice_sign); ?>"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
                 </td>
 
                 <td class="td-amount">
-                    <select name="item_tax_rate_id" name="item_tax_rate_id" class="form-control input-sm">
+                    <select name="item_tax_rate_id" name="item_tax_rate_id" class="form-control input-sm"
+                        <?php if ($invoice->is_read_only == 1) { echo 'disabled="disabled"';} ?>>
                         <option value="0"><?php echo lang('none'); ?></option>
                         <?php foreach ($tax_rates as $tax_rate) { ?>
                             <option value="<?php echo $tax_rate->tax_rate_id; ?>"
@@ -105,10 +111,12 @@
                     </span>
                 </td>
                 <td class="td-icon">
+                    <?php if ($invoice->is_read_only != 1) { ?>
                     <a href="<?php echo site_url('invoices/delete_item/' . $invoice->invoice_id . '/' . $item->item_id); ?>"
                        title="<?php echo lang('delete'); ?>">
                         <i class="fa fa-trash-o no-margin text-danger"></i>
                     </a>
+                    <?php } ?>
                 </td>
             </tr>
         <?php } ?>
@@ -145,7 +153,9 @@
                 <?php
                 if ($invoice_tax_rates) {
                     foreach ($invoice_tax_rates as $invoice_tax_rate) {
-                        echo anchor('invoices/delete_invoice_tax/' . $invoice->invoice_id . '/' . $invoice_tax_rate->invoice_tax_rate_id, '<i class="fa fa-trash-o"></i>');
+                        if ($invoice->is_read_only != 1) {
+                            echo anchor('invoices/delete_invoice_tax/' . $invoice->invoice_id . '/' . $invoice_tax_rate->invoice_tax_rate_id, '<i class="fa fa-trash-o"></i>');
+                        }
                         echo ' ' . $invoice_tax_rate->invoice_tax_rate_name . ' ' . $invoice_tax_rate->invoice_tax_rate_percent; ?>%:
                         <span class="amount">
                             <?php echo format_currency($invoice_tax_rate->invoice_tax_rate_amount * $invoice->invoice_sign); ?>
